jewelry-web

+++ jewelry-server/src/main/java/ro/ausy/jewelry/server/dao/UserRoleDao.java
@@ -6,18 +6,8 @@ import ro.ausy.jewelry.commons.dto.UserRoleDTO;
 
 public interface UserRoleDao {
 	
-	/**
-	 * @return
-	 * This method will return an UserRole by ID.
-	 * @param userRoleId
-	 * userRoleId.
-	 */
-	UserRoleDTO getUserRoleById(long userRoleId);
+	UserRoleDTO getUserRoleById(int userRoleId);
 	
-	/**
-	 * @return
-	 * This method will return all usersRole.
-	 */
 	List<UserRoleDTO> getAllUserRole();
 
 }
